Use md5.js instead of crypto
The version before this pulls in crypto and only uses it for MD5
hashing, which is fine, but it balloons the size of the mii-editor
bundle to 904 KB. With this change, it becomes more like 293 KB.

Commit was amended on 2024-04-19 due to accidentally including "md5-js"
in package.json instead of "md5.js". Oops!